OxxiusCombiner: Update by Tristan Martinez, Oxxius
Replace all .h and .cpp with new files from Tristan.
Update .vcxproj and Makefile.am accordingly.
Add previously missing .vcxproj.filters.
