+++ DeviceAdapters/OxxiusCombiner/Makefile.am
@@ -1,6 +1,16 @@
 AM_CXXFLAGS = $(MMDEVAPI_CXXFLAGS)
 deviceadapter_LTLIBRARIES = libmmgr_dal_OxxiusCombiner.la
-libmmgr_dal_OxxiusCombiner_la_SOURCES = OnBoardHW.cpp OnBoardHW.h Oxxius_combiner.cpp Oxxius_combiner.h
+libmmgr_dal_OxxiusCombiner_la_SOURCES = \
+					Cobolt08_01.cpp \
+					CoherentObis.cpp \
+					GenericLaser.cpp \
+					OxxiusCombinerHub.cpp \
+					OxxiusFlipMirror.cpp \
+					OxxiusLBX.cpp \
+					OxxiusLCX.cpp \
+					OxxiusMDual.cpp \
+					OxxiusShutter.cpp
+
 libmmgr_dal_OxxiusCombiner_la_LIBADD = $(MMDEVAPI_LIBADD)
 libmmgr_dal_OxxiusCombiner_la_LDFLAGS = $(MMDEVAPI_LDFLAGS)
