+++ DeviceAdapters/OxxiusCombiner/OxxiusCombiner.vcxproj.filters
@@ -1,0 +1,69 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <Filter Include="Header Files">
+      <UniqueIdentifier>{15751bdb-12d5-452c-8197-c6762010b6b3}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files">
+      <UniqueIdentifier>{d5eef910-aac2-4511-8c33-ad79c7499d4d}</UniqueIdentifier>
+    </Filter>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Cobolt08_01.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="CoherentObis.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="GenericLaser.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="OxxiusCombinerHub.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="OxxiusFlipMirror.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="OxxiusLBX.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="OxxiusLCX.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="OxxiusMDual.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="OxxiusShutter.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="Cobolt08_01.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="CoherentObis.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="GenericLaser.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="OxxiusCombinerHub.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="OxxiusFlipMirror.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="OxxiusLBX.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="OxxiusLCX.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="OxxiusMDual.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="OxxiusShutter.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+  </ItemGroup>
+</Project>
